% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfgCost.R
\name{mfgCost_delete}
\alias{mfgCost_delete}
\title{删除月份数据}
\usage{
mfgCost_delete(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FYear = 2022,
  FPeriod = 9
)
}
\arguments{
\item{dms_token}{口令}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
删除月份数据
}
\examples{
mfgCost_delete()
}
