% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfgCost.R
\name{mfgCost_Read}
\alias{mfgCost_Read}
\title{读取模板成本}
\usage{
mfgCost_Read(
  file_name = "data-raw/产品生产成本模板.xlsx",
  sheet = "data",
  FYear = 2022,
  FPeriod = 9,
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{file_name}{文件名}

\item{sheet}{页签名}

\item{FYear}{年份}

\item{FPeriod}{月份}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
读取模板成本
}
\examples{
mfgCost_Read()
}
