% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mfgCost.R
\name{mfgCost_Server}
\alias{mfgCost_Server}
\title{管理价格查询}
\usage{
mfgCost_Server(input, output, session, dms_token)
}
\arguments{
\item{input}{输入}

\item{output}{输出}

\item{session}{会话}

\item{dms_token}{口令}
}
\value{
返回值
}
\description{
管理价格查询
}
\examples{
mfgCost_Server()
}
