% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngrCost.R
\name{mngrCost_Read}
\alias{mngrCost_Read}
\title{读取模板成本}
\usage{
mngrCost_Read(
  file_name = "data-raw/白标售价及管理成本20220903_upload_reviewed.xlsx",
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3"
)
}
\arguments{
\item{file_name}{文件名}
}
\value{
返回值
}
\description{
读取模板成本
}
