% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mngrCost.R
\name{mngrCost_SyncLastOne}
\alias{mngrCost_SyncLastOne}
\title{同步上一期数据}
\usage{
mngrCost_SyncLastOne(
  dms_token = "9B6F803F-9D37-41A2-BDA0-70A7179AF0F3",
  FYear = 2022,
  FPeriod = 9
)
}
\arguments{
\item{dms_token}{口令}

\item{FYear}{年份}

\item{FPeriod}{月份}
}
\value{
返回值
}
\description{
同步上一期数据
}
\examples{
mngrCost_SyncLastOne()
}
